% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_CreateTimeDividers.R
\name{CreateTimeDividers}
\alias{CreateTimeDividers}
\title{Create period divides as a function of time from start of toxicity window}
\usage{
CreateTimeDividers(rt, timeBoundaries, labelUnits = "days")
}
\arguments{
\item{rt}{Object of class robustToxicitiesClass}

\item{timeBoundaries}{a numeric vector of times from the start of toxicity window}

\item{labelUnits}{days, weeks or months used to automatically generate labels in the form of from-to.}
}
\description{
Takes a set of numeric time boundaries from baseline and creates the corresponding columns in patientData, periodDividerCols and periodDividerLabels. This returns the updated robustToxicitiesClass object.
}
