% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_toxicityOptions.R
\name{DefaultToxicityOptions}
\alias{DefaultToxicityOptions}
\title{Default \code{\link{toxicityOptions}}.}
\usage{
DefaultToxicityOptions()
}
\description{
Default \code{\link{toxicityOptions}}.
}
