% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxTable_FlexTable_wrappers.R
\name{FT_ToxTable}
\alias{FT_ToxTable}
\alias{FT_ToxTable_summary}
\alias{FT_ToxTable_cycle}
\alias{FT_ToxTable_category}
\title{FlexTable wrapper for toxicity tables}
\usage{
FT_ToxTable_summary(rt, tble = NULL)

FT_ToxTable_cycle(rt, cycles = "all", tble = NULL)

FT_ToxTable_category(rt, cycles = "all", tble = NULL)
}
\arguments{
\item{rt}{robustToxicity object.}

\item{tble}{Optionally pass a pre computed table into the wrapper instead of computing it here. Note if columns don't match what is expected this is likely to fail. May be usefull for dropping some rows, or saving time rerunning the table generator function for large data sets.

This functions return the ConstructFlexTable object from the prettyTables package. To create the FlexTable you can call object$GetTable() on the returned object. This can then be added to a word document using \code{\link{addFlexTable}} from the ReporteRs package.}

\item{cycles}{The cycle column names, or index in rt@periodDividerCols of the cycles to tabulate. May also be "all" to use all cycles}
}
\description{
These functions act as wrapper functions to create ConstructFlexTable objects.
}
\section{Functions}{
\itemize{
\item \code{FT_ToxTable_summary}: A wrapper for ToxTable_cycle

\item \code{FT_ToxTable_cycle}: A wrapper for ToxTable_cycle

\item \code{FT_ToxTable_category}: A wrapper for ToxTable_category
}}

\examples{
# Patient Level Data
data("rt_patientData")
# Toxicity Level Data
data("rt_toxicityData")


# Run the setup command passing in all the column names.
rt = SetupRobustToxicities(
  toxData = rt_toxicityData,
  patientData = rt_patientData,
  patidCol = "patientNo", treatmentCol = "Treatment",
  toxCategoryCol = "category", toxNameCol = "toxicity",
  toxGradeCol = "grade", dateOfStartOfToxWindow = "Registration_date",
  dateOfStartTox = "ae_onset_date", dateOfEndTox = "ae_resolve_date",
  dateOfEndOfToxWindow = "end_of_assessment_date",
  periodDividerCols = c("Registration_date", "Cycle_1_date","Cycle_2_date",
                        "Cycle_3_date", "Cycle_4_date", "Cycle_5_date", "Cycle_6_date"),
  periodDividerLabels = c("Pre treatment", "Cycle 1","Cycle 2",
                          "Cycle 3","Cycle 4","Cycle 5", "Cycle 6"),
  treatmentCodes = NULL, treatmentLabels = NULL, options = NULL)

# Look for queries. Note: must be called before running any
# of the functions on this class.
rt = QueryRobustToxicities(rt)


##########################################################
# flexTable summary
ft = FT_ToxTable_summary(rt)
# returns an object of class ConstructFlexTable for additional formatting and editing
ft$GetTable()


##########################################################
# flexTable worst grade of each adverse event by patient
ft = FT_ToxTable_cycle(rt)
# returns an object of class ConstructFlexTable for additional formatting and editing
ft$GetTable()

##########################################################
# flexTable worst grade of each category by patient
ft = FT_ToxTable_category(rt)
# returns an object of class ConstructFlexTable for additional formatting and editing
ft$GetTable()

##########################################################
# The table generated by ToxTable_summary, ToxTable_cycle or ToxTable_category may also
# be passed as an argument. This provides a convinent oportunity to edit the data
tble = ToxTable_summary(rt)
# Edits to tble can go here
ft = FT_ToxTable_summary(rt, tble = tble)
# returns an object of class ConstructFlexTable for additional formatting and editing
ft$GetTable()
}
\seealso{
\code{\link{ToxTable_summary}}, \code{\link{ToxTable_cycle}}, \code{\link{ToxTable_category}}
}
