% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameDatabase.R
\name{nameDatabase}
\alias{nameDatabase}
\title{Renaming function to generate the correct names for the robustToxicities class}
\usage{
nameDatabase(data, patid = NULL, treatment = NULL, ae_term = NULL,
  ae_system = NULL, ae_grade = NULL, ae_start_date = NULL,
  ae_end_date = NULL, ae_cont_end_study = NULL,
  date_stopped_treatment = NULL, date_end_assessment = NULL,
  dateColumnNames = NULL, patientInCycle = NULL, occurInCycle = NULL,
  ass_TRUE = NULL)
}
\arguments{
\item{data}{The toxicity database}

\item{patid}{Patient ID}

\item{treatment}{Treatment arms}

\item{ae_term}{The adverse event name / term}

\item{ae_system}{The adverse event system}

\item{ae_grade}{The adverse event grade}

\item{ae_start_date}{The adverse event start date (time based only)}

\item{ae_end_date}{The adverse event end date (time based only)}

\item{ae_cont_end_study}{Was the adverse event coninuing at the end of the study (time based only)}

\item{date_stopped_treatment}{Date the patients stopped treatment (time based only)}

\item{date_end_assessment}{Date the patients stop being assessed for toxicity (time based only)}

\item{dateColumnNames}{The start date of each cycle / time period (time based only)}

\item{patientInCycle}{Was a patient enrolled for this cycle / time period (cycle based, can be auto genrated for time based data)}

\item{occurInCycle}{Was the toxicity present for this cycle / time period (cycle based, can be auto genrated for time based data)}

\item{ass_TRUE}{An indicator to determine if the toxicity should be counted in the analysis}
}
\description{
This function provides an efficient way of renaming the column names in the toxicity database to be compliant with the function \code{\link{robustToxicities}}.
}

