% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_toxicityTime.R
\name{plot_toxicityTime}
\alias{plot_toxicityTime}
\title{Plot toxicities over time}
\usage{
plot_toxicityTime(toxDB, patid = character(0), dayRange = c(-7, 60),
  omit = 1, plot = TRUE, plotCycleLength = 21)
}
\arguments{
\item{toxDB}{Main toxicity database.}

\item{patid}{Build in patient subsetting by a list if not ""}

\item{dayRange}{The range of time in days to display on the graph}

\item{plot}{TRUE/FALSE plot the graph.}

\item{plotCycleLength}{}
}
\value{
This plot function return the number of row of unique toxicities * patients. This assists in computing optimal size for saved graphs.
}
\description{
This function plots all the toxicities provided over time
}

