% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toxicity_prepareToxicity.R
\name{prepareToxicity}
\alias{prepareToxicity}
\title{Prepare toxicities}
\usage{
prepareToxicity(toxDB)
}
\arguments{
\item{toxDB}{Main toxicity database.}
}
\description{
Performs a number of data cleaning actions to ensure the data is appropriate for tabulation and plotting
}

