% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_toxTable_cycle.R
\name{print_toxTable_cycle}
\alias{print_toxTable_cycle}
\title{Format the toxTable_summary to an output medium}
\usage{
print_toxTable_cycle(toxDB, cycles, printMethod = "print", rtfDoc = NULL)
}
\arguments{
\item{cycles}{The cycle or cycles to generate the table for}

\item{printMethod}{One of "print" "rtf" or "latex"}

\item{rtfDoc}{the name of the rtf document to output to}
}
\description{
Format the toxTable_summary to an output medium
}

