% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_toxTable_summary.R
\name{print_toxTable_summary}
\alias{print_toxTable_summary}
\title{Format the toxTable_summary}
\usage{
print_toxTable_summary(toxDB, printMethod = "table")
}
\arguments{
\item{printMethod}{One of "table" or "latex"}
}
\description{
Returns a clean version of toxTable_summary.
Also includes an option to output a version for latex.
}
\details{
The latex option details requires you to use the \code{array} and \code{multirow} packages in the .tex file using \code{\\usepackage{array, multirow}}.
}

