% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryRobustToxicities.R
\name{QueryRobustToxicities}
\alias{QueryRobustToxicities}
\title{Query RobustToxicities}
\usage{
rt = QueryRobustToxicities(rt)
}
\arguments{
\item{rt}{The robustToxicitiesClass object}
}
\value{
An S4 object of class robustToxicitiesClass. The queries are stored as a data.frame in the slot queries (rt@queires)
}
\description{
A function which checks the provided data
}
