% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_robustToxicities.R
\docType{class}
\name{robustToxicities-class}
\alias{.robustToxicities}
\alias{robustToxicities-class}
\title{The robustToxicities class}
\description{
This is core object of this package. This object stores the original dataset as well as the automatically cleaned dataset and a list of notes and queries generated when cleaning the dataset. A list of options is also provided to store plot and tabulation options and provide additional metadata. Finally treatment and cycle labels are also required.
}
\section{Slots}{

\describe{
\item{\code{data}}{The original dataset}

\item{\code{cleanData}}{The cleaned dataset}

\item{\code{treatmentLabels}}{A vector of treatment labels}

\item{\code{cycleLabels}}{A vector of cycle / time perior labels}

\item{\code{options}}{An s4 object of class \code{\link{toxicityOptions}} containing options and metadata for the files.}
}}

