% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_robustToxicities.R
\name{robustToxicities}
\alias{robustToxicities}
\title{The robustToxicities generator}
\usage{
robustToxicities(data, cycleLabels, options, treatmentLabels = NULL)
}
\arguments{
\item{data}{The data set to use}

\item{cycleLabels}{Labels for each cycle or time period to using in tables}

\item{options}{A set of options to use when tabulating and plotting}

\item{treatmentLabels}{What each treatment arm should be called in plots and tables}
}
\description{
The robustToxicities generator
}

