% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_HelpFile.R
\docType{data}
\name{rt_patientData}
\alias{rt_patientData}
\alias{rt_toxicityData}
\title{Toxicity data}
\format{Two data.frames which are linked by patientNo. rt_patientData contains data for 7 patients with one row per patient. rt_toxicityData contains 29 toxicities for the patients in rt_patientData.}
\usage{
data("rt_patientData")
data("rt_toxicityData")
}
\description{
An example dataset of dummy data used to create examples for this package.
}
\keyword{datasets}
