% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_robustToxicities.R
\name{SetupRobustToxicities}
\alias{SetupRobustToxicities}
\title{The robustToxicities generator}
\usage{
SetupRobustToxicities(toxData, patientData, patidCol, treatmentCol = NULL,
  toxCategoryCol, toxNameCol, toxGradeCol, dateOfStartOfToxWindow,
  dateOfStartTox, dateOfEndTox, dateOfEndOfToxWindow,
  periodDividerCols = character(0), periodDividerLabels = character(0),
  treatmentCodes = NULL, treatmentLabels = NULL, options = NULL)
}
\arguments{
\item{toxData}{The toxicity level data set}

\item{patientData}{The patient level data}

\item{patidCol}{Column name for the participant identifier}

\item{treatmentCol}{Column name for the treatment. Will be created if not provided}

\item{toxCategoryCol}{Column name for aderse event category}

\item{toxNameCol}{Column name for adverse event name}

\item{toxGradeCol}{Column name for the adverse event grade}

\item{dateOfStartOfToxWindow}{Column name for date of study entry (eg registration)}

\item{dateOfStartTox}{Column name for date of adverse event start or change in grade}

\item{dateOfEndTox}{Column name for date of adverse event end or change in grade}

\item{dateOfEndOfToxWindow}{Column name for the end of the time window for the particitant to be observed for toxicities (optional)}

\item{periodDividerCols}{Column names for date dividing times into periods or cycles (optional)}

\item{periodDividerLabels}{Display names for data periodDividerCols
param treatmentLabels A vector of treatment labels}

\item{options}{Optional. An object of class toxicityOptions. The easiest place to start is with \code{DefaultToxicityOptions()}.}
}
\description{
The robustToxicities generator
}
