% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worstGradeByPatient.R
\name{table_tox_categories}
\alias{table_tox_categories}
\title{Generate a prettyTable of worst toxicity by category.}
\usage{
table_tox_categories(data, categoryList = "all", strata = NULL,
  strata.names = NULL, strata.count = TRUE, overall = TRUE, count = "n",
  round = 3)
}
\arguments{
\item{data}{A data.frame This should be the output from \code{\link{worstGradeByPatientSystem}}}

\item{categoryList}{A list of categories. Default is all. A subset is selected by changing this value.}

\item{strata}{A string of the column to used to stratify on}

\item{strata.names}{Otional. A named vector of alternative strata names}

\item{strata.count}{TRUE/FALSE for displaying strata counts at the top of each column}

\item{overall}{TRUE/FALSE for including an overall column}

\item{count}{"n","miss" or "none" providing the counts, missing values or omitting for each column for numeric variables}

\item{round}{A value or vector for the number of significant figures to report the data to

categoryList can be replaced by other column names if the standard categorisation is not being used. For more flexibility you can view the code by typing table_tox_categories.

Available methods and values for \strong{Type}:
\tabular{cc}{ "miqr" \tab median (Q25,Q75) \cr "miqrr" \tab median (Q25,Q75)[min,max] \cr "mrng" \tab median (Q0,Q100) \cr "avsd" \tab mean (sd) \cr "avci" \tab mean (confidence interval) \cr "st" \tab count \cr "str" \tab count/total \cr "stp" \tab count (percent) \cr "strp" \tab count/total (percent)
}}

\item{type}{A vector of the method to summarise the var}
}
\value{
Returns a data.frame
}
\description{
This is a wrapper for the table_values function in prettyTables for worst grade by system.
}

