% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxData.R
\docType{data}
\name{toxicityDataTime}
\alias{toxicityDataTime}
\title{Toxicity data time}
\format{A data frame with 29 rows and 20 variables}
\usage{
data(toxicityDataTime)

data(toxicityDataCycle)
}
\description{
A time based data set of toxicities. Used as one of the key examples for the robustToxicities pacakge

A time based data set of toxicities. Used as one of the key examples for the robustToxicities pacakge
}
\keyword{datasets}

