% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_toxicityOptions.R
\docType{class}
\name{toxicityOptions-class}
\alias{toxicityOptions-class}
\alias{.toxicityOptions}
\alias{DefaultToxicityOptions}
\title{Toxicity Options}
\usage{
DefaultToxicityOptions()
}
\description{
An object containing all the key options for creating the toxicity tables.
This is a key slot in the \code{\link{robustToxicities-class}}.

Default toxicity options generator, requiring metadata only.
}
\section{Slots}{

\describe{
\item{\code{displayNotes}}{A logical value used by \code{\link{robustToxicities}} to determine whether to print note or note}

\item{\code{toxTable_cycle_tabulationMethod}}{One of "worst" or "all" determining if all toxicity changes are counted or only the worst reported grade in a time period}

\item{\code{toxTable_tabulationPercent}}{A logical value used to determine if toxicity tables should report counts (FALSE) or percentages (TRUE)}

\item{\code{toxTable_tabulationZeros}}{A logical value used to determine if zeros should be included, default TRUE}

\item{\code{toxTable_cumulativeGrades}}{A logical value used to determine whether toxicity grades should be reported cumulatively or not}

\item{\code{toxTable_discardToxAtStudyEntry}}{A logical value used to determine if toxicities reported at baseline should be reported or not}

\item{\code{toxTable_mergeGrades}}{Grades to merge in the \code{\link{print_toxTable_cycle}}. Columns are seperated by "|" and merged values are seperated by ",". "n"}

\item{\code{toxTable_cycle_toxicityOrder}}{What order should the data be returned in. "c" ordered by categories and then adverse events. "a" ordered by adverse events. "n" ordered by number of adverse events. The n option can be followed by a number to denote the minimum grade to use for sorting. e.g. "n3" will order by grades 3-5 and then 1-5 for ties within grades 3-5.}
}}

