% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxPlot.R
\name{toxPlot}
\alias{toxPlot}
\title{Plot toxicities}
\usage{
toxPlot(toxDB, rowID_range = NULL, plot = TRUE)
}
\arguments{
\item{toxDB}{an object of class robustToxicities}

\item{rowID_range}{optional, a length 2 vector detailing the minimum and maximum row to plot}

\item{plot}{whether to plot the graph or return the number of rows to plot}
}
\value{
This plot function return the number of row of unique toxicities * patients. This assists in computing optimal size for saved graphs.
}
\description{
This function plots all the toxicities provided over time
}

