% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxPlot_byCycle.R
\name{ToxPlot_byCycle}
\alias{ToxPlot_byCycle}
\title{Summary plot of toxicities by cycle}
\usage{
ToxPlot_byCycle(rt, gradeRequired = 1, colors = c("blue", "red"),
  tableSpace = 0.1, las = 1, legendPosition = "right", add = FALSE)
}
\arguments{
\item{rt}{An object of class robustToxicities}

\item{gradeRequired}{Only include adverse events with at least this grade}

\item{tableSpace}{A parameter to assist in vertical row spacing the table appropriately}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. {0: always parallel to the axis}, {1: always horizontal [default]}, {2: always perpendicular to the axis}, {3: always vertical}}

\item{legendPosition}{The location to place the legend see \code{\link{legend}} for details}

\item{add}{TRUE/FALSE whether to add to an existing plot or start a new one}

\item{timeBoundaries}{A vector of times from the dateOfStartOfToxWindow}

\item{xlab}{The xaxis label. "days","weeks" and "months" are converted to "Time from registration (days)" etc.}

\item{col}{A vector of colours to plot each arm with}
}
\description{
This plot summarises the proportion of patients having an adverse event in each time period as defined by periodDividerCols in the \code{\link{robustToxicitiesClass}}.
}
\examples{
# Patient Level Data
data("rt_patientData")
# Toxicity Level Data
data("rt_toxicityData")


# Run the setup command passing in all the column names.
rt = SetupRobustToxicities(
  toxData = rt_toxicityData,
  patientData = rt_patientData,
  patidCol = "patientNo", treatmentCol = "Treatment",
  toxCategoryCol = "category", toxNameCol = "toxicity",
  toxGradeCol = "grade", dateOfStartOfToxWindow = "Registration_date",
  dateOfStartTox = "ae_onset_date", dateOfEndTox = "ae_resolve_date",
  dateOfEndOfToxWindow = "end_of_assessment_date",
  periodDividerCols = c("Registration_date", "Cycle_1_date","Cycle_2_date",
                        "Cycle_3_date", "Cycle_4_date", "Cycle_5_date", "Cycle_6_date"),
  periodDividerLabels = c("Pre treatment", "Cycle 1","Cycle 2",
                          "Cycle 3","Cycle 4","Cycle 5", "Cycle 6"),
  treatmentCodes = NULL, treatmentLabels = NULL, options = NULL)

# Look for queries. Note: must be called before running any
# of the functions on this class.
rt = QueryRobustToxicities(rt)

##########################################################
par(mar = c(4,6,3,2))
ToxPlot_byCycle(rt)

##########################################################
# subset to a specific set of adverse events
rt@toxData$ass_TRUE = rt@toxData$toxicity == "Sore Throat"
par(mar = c(4,6,3,2))
ToxPlot_byCycle(rt)
}
\seealso{
\code{\link{ToxPlot_byPatient}}, \code{\link{ToxPlot_byToxicity}}, \code{\link{ToxPlot_byTime}}
}
