% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxPlot_byTime.R
\name{ToxPlot_byTime}
\alias{ToxPlot_byTime}
\title{Summary plot of toxicities over time}
\usage{
ToxPlot_byTime(rt, gradeRequired = 1, timeBoundaries, xlab = "days",
  col = c("blue", "red"), tableSpace = 0.1, las = 1,
  legendPosition = "right", add = FALSE)
}
\arguments{
\item{rt}{An object of class robustToxicities}

\item{gradeRequired}{Only include adverse events with at least this grade}

\item{timeBoundaries}{A vector of times from the dateOfStartOfToxWindow}

\item{xlab}{The xaxis label. "days","weeks" and "months" are converted to "Time from registration (days)" etc.}

\item{col}{A vector of colours to plot each arm with}

\item{tableSpace}{A parameter to assist in vertical row spacing the table appropriately}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. {0: always parallel to the axis}, {1: always horizontal [default]}, {2: always perpendicular to the axis}, {3: always vertical}}

\item{legendPosition}{The location to place the legend see \code{\link{legend}} for details}

\item{add}{TRUE/FALSE whether to add to an existing plot or start a new one}
}
\description{
This is a wrapper function for \code{\link{ToxPlot_byCycle}} which takes timeBoundaries.
}
\seealso{
\code{\link{ToxPlot_byPatient}}, \code{\link{ToxPlot_byToxicity}}, \code{\link{ToxPlot_byCycle}}
}
