% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_causalityInfo.R
\name{ToxPlot_causalityInfo}
\alias{ToxPlot_causalityInfo}
\alias{causalityInfo,}
\alias{causalityInfo-class}
\title{causalityInfo class}
\usage{
ToxPlot_causalityInfo(columns, names = character(0), width = 1.5,
  pch = c(NA, NA, 4, 8, 16), cex = 1, col = 1,
  labels = c("Possibly related", "Probably related", "Definitely related"))
}
\arguments{
\item{columns}{A vector of column names for toxData}

\item{names}{A short identifier to place at the top of each column}

\item{width}{The width to provide on the plot for each causality. This is on the scale of days on the plot}

\item{pch}{The pch symbol to use for each level of causality}

\item{cex}{The size of each symbol}

\item{col}{The colour to use for each level of causality}

\item{labels}{Labels for the non NA pch levels, used in the legend}
}
\value{
An object of class causalityInfo containing the same slots as paramters taken by ToxPlot_causalityInfo
}
\description{
Stores causality data to pass to toxPlot_byToxicity in the causality parameter. The columns containing the data in toxData should hold numeric data.
}
\details{
The default values expect a number between 1 and 5 and plot a symbol for values 3,4 and 5. This can be changed using pch.
}
