% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_eventInfo.R
\name{ToxPlot_eventInfo}
\alias{ToxPlot_eventInfo}
\alias{eventInfo,}
\alias{eventInfo-class}
\title{eventInfo class}
\usage{
ToxPlot_eventInfo(columns, labels = columns, lwd = 1.5, col = 1)
}
\arguments{
\item{columns}{A vector of column names for toxData}

\item{labels}{A short name for each event}

\item{lwd}{Line width}

\item{col}{The colour to use for each event}
}
\value{
An object of class eventInfo containing the same slots as paramters taken by ToxPlot_causalityInfo
}
\description{
Stores event data to pass to toxPlot_byToxicity in the causality parameter. The columns containing the data in patientData should hold date data or \code{\link{difftime}} data in days.
}
