% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxPlot_output.R
\name{toxPlot_output}
\alias{toxPlot_output}
\title{Toxixicity plot output to png and pdf}
\usage{
toxPlot_output(toxDB, pageLimit = 20, type = "pdf", width = 9,
  height = NULL)
}
\arguments{
\item{toxDB}{The robustToxicities object to plot}

\item{pageLimit}{The max number of row to add per page}

\item{type}{One of "pdf" or "png"}

\item{width}{The width of the plot output. It is recommended to use 9 for pdf and 1100 for png}

\item{height}{The height of the plot output. If se to NULL this is automatically calculated based on the pageSize.}
}
\description{
This function acts as a wrapper to output png and pdf plots for toxPlot.
}

