% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_all.R
\name{toxTable_all}
\alias{toxTable_all}
\title{Table: all toxicity episodes}
\usage{
toxTable_all(toxDB, treatments)
}
\arguments{
\item{toxDB}{Main toxicity database.}
}
\description{
Generates a table of all toxicities
}

