% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_cycle.R
\name{ToxTable_cycle}
\alias{ToxTable_cycle}
\title{Toxicity tables by cycle/s}
\usage{
ToxTable_cycle(rt, cycles = "all")
}
\arguments{
\item{rt}{an object of class robustToxicities}

\item{cycles}{The cycle column names, or index in rt@periodDividerCols of the cycles to tabulate. May also be "all" to use all cycles}
}
\description{
Returns a toxicity table with the requested data according to the ass_TRUE column, for the cycles requested.
}
