% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_summary.R
\name{toxTable_summary}
\alias{toxTable_summary}
\title{Table: Summary of worst toxicity by cycle}
\usage{
toxTable_summary(toxDB)
}
\arguments{
\item{toxDB}{Main toxicity database.}
}
\description{
Summarises toxicities by treatment and cycle
}

