% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_worst.R
\name{toxTable_worst}
\alias{toxTable_worst}
\title{Table: Worst toxicity}
\usage{
toxTable_worst(toxDB, treatments)
}
\arguments{
\item{toxDB}{Main toxicity database.}

\item{treatments}{vector of treatments}
}
\description{
Table: Worst toxicity
}

