% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_worstInCycle.R
\name{toxTable_worstInCycle}
\alias{toxTable_worstInCycle}
\title{Table: Toxicity worst in cycle}
\usage{
toxTable_worstInCycle(toxDB, cycles, worst = "worst")
}
\arguments{
\item{toxDB}{Main toxicity database.}
}
\description{
Generates a table of all toxicities for a certain cycle or cycles. There are options to report all changes in toxicities or just the worst per patient. CTCAE grades and categories can be merged for more sucinct tables.
}

