% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_worstSummary.R
\name{toxTable_worstSummary}
\alias{toxTable_worstSummary}
\title{Table: Summary of worst toxicity by cycle}
\usage{
toxTable_worstSummary(toxDB, cycleMerge = character(0),
  columnMerge = character(0))
}
\arguments{
\item{toxDB}{Main toxicity database.}
}
\description{
Summarises toxicities by treatment and cycle
}

