% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worstGradeByPatient.R
\name{worstGradeByPatient}
\alias{worstGradeByPatient}
\alias{worstGradeByPatientSystem}
\title{Generate worst grade by patient}
\usage{
worstGradeByPatient(rt, toxData = NULL)

worstGradeByPatientSystem(rt, categoryList = NULL)
}
\arguments{
\item{rt}{an object of class robustToxicities}

\item{toxData}{A data.frame subset of rt@toxData if not all data should be used}
}
\value{
Returns a data.frame
}
\description{
Returns a patient level data.frame containing the worst grade for each patient. The ass_TRUE is used as a filter. Optionally a subset of rt@toxData can be passed in as the second variable. In this case ass_TRUE is still used as a filter.
}
\details{
worstGradeByPatientSystem does the same thing but wraps over all categories.
}
