% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxTable_FlexTable_wrappers.R
\name{FT_ToxTable}
\alias{FT_ToxTable}
\alias{FT_ToxTable_summary}
\alias{FT_ToxTable_cycle}
\alias{FT_ToxTable_category}
\title{FlexTable wrapper for toxicity tables}
\usage{
FT_ToxTable_summary(rt, tble = NULL)

FT_ToxTable_cycle(rt, cycles = "all", tble = NULL)

FT_ToxTable_category(rt, cycles = "all", tble = NULL)
}
\arguments{
\item{rt}{robustToxicity object.}

\item{tble}{Optionally pass a pre computed table into the wrapper instead of computing it here. Note if columns don't match what is expected this is likely to fail. May be usefull for dropping some rows, or saving time rerunning the table generator function for large data sets.

This functions return the ConstructFlexTable object from the prettyTables package. To create the FlexTable you can call object$GetTable() on the returned object. This can then be added to a word document using \code{\link{addFlexTable}} from the ReporteRs package.}

\item{cycles}{The cycle column names, or index in rt@periodDividerCols of the cycles to tabulate. May also be "all" to use all cycles}
}
\description{
These functions act as wrapper functions to create ConstructFlexTable objects.
}
\section{Functions}{
\itemize{
\item \code{FT_ToxTable_summary}: A wrapper for ToxTable_cycle

\item \code{FT_ToxTable_cycle}: A wrapper for ToxTable_cycle

\item \code{FT_ToxTable_category}: A wrapper for ToxTable_category
}}

