% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_toxTable_summary.R
\name{print_toxTable_summary}
\alias{print_toxTable_summary}
\title{Format the toxTable_summary to an output medium}
\usage{
print_toxTable_summary(toxDB, printMethod)
}
\arguments{
\item{toxDB}{Main toxicity database.}

\item{printMethod}{One of "rft" or "latex"}
}
\description{
Format the toxTable_summary to an output medium
}

