% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_toxTable_summary.R
\name{print_toxTable_summary}
\alias{print_toxTable_summary}
\title{Toxicity table summary}
\usage{
print_toxTable_summary(toxDB, printMethod = "table")
}
\arguments{
\item{toxDB}{an object of class robustToxicities}

\item{printMethod}{One of "print" "table" or "latex"}
}
\description{
Returns a summary toxicity table with the requested data according to the ass_TRUE column.
}
\details{
The latex option details requires you to use the \code{array} and \code{multirow} packages in the .tex file using \code{\\usepackage{array, multirow}}.
}

