% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryRobustToxicities.R
\name{QueryRobustToxicities}
\alias{QueryRobustToxicities}
\title{Query RobustToxicities}
\usage{
QueryRobustToxicities(rt)
}
\arguments{
\item{rt}{The robustToxicitiesClass object}
}
\description{
A function which checks the provided data
}
