\name{robustToxicities-package}
\alias{robustToxicities}
\docType{package}
\title{Toxicity tabulation and plots}
\description{
This shiny application collates toxicities into publication ready tables both in csv format and word document tables. There is also a plot showing the toxicities over time.
}
\details{
\tabular{ll}{
Package: \tab robustToxicities\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2016-03-11\cr
License: \tab GPL-3\cr
}
This package contains a shiny application to create customised graphs and tables of trial toxicity data.

List of user apps:\cr
\code{\link{shiny_robustToxicities}}\cr
\code{\link{shiny_robustToxicities.log}}
}
\author{
Peter Dutton

Maintainer: Peter Dutton <dutton.peter@gmail.com>
}
\references{
None
}
\keyword{ package}
\seealso{
}
\examples{
}
