\name{robustToxicities-package}
\alias{robustToxicities}
\docType{package}
\title{Toxicity tabulation and plots}
\description{
This shiny application collates toxicities into publication ready tables both in csv format and word document tables. There is also a plot showing the toxicities over time.
}
\details{
\tabular{ll}{
Package: \tab robustToxicities\cr
Type: \tab Package\cr
Version: \tab 1.02\cr
Date: \tab 2017-08-14\cr
License: \tab GPL-3\cr
}
}
