% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_toxicityOptions.R
\docType{class}
\name{toxicityOptions-class}
\alias{defaultToxicityOptions}
\alias{toxicityOptions-class}
\title{Toxicity Options}
\usage{
defaultToxicityOptions(trialName, folderPath = NULL, fileName = "",
  timeType = "time")
}
\arguments{
\item{trialName}{Trial or study Name}

\item{folderPath}{Path to the folder containing the data}

\item{fileName}{Name of the file containing the data}

\item{timeType}{One of "Time" or "Cycle" to denote whether the data is primarily time based or only cycle / time period based}
}
\description{
An object containing all the key options for creating the toxicity tables.
This is a key slot in the \code{\link{robustToxicities-class}}.

Default toxicity options generator, requiring metadata only.
}
\section{Slots}{

\describe{
\item{\code{trialName}}{Trial Name (character string)}

\item{\code{folderPath}}{Reference to the folder the data file is stored in (character string)}

\item{\code{fileName}}{Reference to the stored data (character string)}

\item{\code{timeType}}{One of "Time" or "Cycle" to denote whether the data is primarily time based or only cycle / time period based}

\item{\code{displayNotes}}{A logical value used by \code{\link{robustToxicities}} to determine whether to print note or note}

\item{\code{tabulationMethod}}{One of "worst" or "all" determining if all toxicity changes are counted or only the worst reported grade in a time period}

\item{\code{cumulativeGrades}}{A logical value used to determine whether toxicity grades should be reported cumulatively or not}

\item{\code{discardBaseline}}{A logical value used to determine if toxicities reported at baseline should be reported or not}

\item{\code{plotStartTreatment}}{The column name of the treatment start date used in plotting (character string)}

\item{\code{plotxMin}}{minimum x axis limit}

\item{\code{plotxMax}}{maximum x axis limit}

\item{\code{plotCycleLength}}{optional cycle length value}

\item{\code{plotPxHeight}}{Number of pixels to use to generate plot vertically}

\item{\code{plotPxWidth}}{Number of pixels to use to generate plot horizontally}

\item{\code{sumCycleMerge}}{Cycles to merge in the \code{\link{print_toxTable_summary}}. Use numeric values with | to divide the merged cycles and , to divide cycles in a merge e.g. "1,2|3,4,5" is two merged time periods with the first 2 time periods and the last 3 time periods.}

\item{\code{sumColumnMerge}}{Grades to merge in the \code{\link{print_toxTable_summary}}. Similar syntax to sumCycleMerge}

\item{\code{cycleCycleMerge}}{Cycles to merge in the \code{\link{print_toxTable_cycle}}. Similar syntax to sumCycleMerge}

\item{\code{cycleColumnMerge}}{Grades to merge in the \code{\link{print_toxTable_cycle}}. Similar syntax to sumCycleMerge}

\item{\code{cycleCategoryMerge}}{A list of categories to collapse down to one row in the \code{\link{print_toxTable_cycle}}.}
}}

