% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxPlot_byToxicity.R
\name{toxPlot_byToxicity}
\alias{toxPlot_byToxicity}
\title{Plot toxicities}
\usage{
toxPlot_byToxicity(rt, rowID_range = NULL, plot = TRUE,
  plotLeftSideOption = "treatment", xlim = c(-7, 60),
  plotCycleLength = 21, plotCycles = 6, plotXLegendScale = "days")
}
\arguments{
\item{rt}{an object of class robustToxicities}

\item{rowID_range}{optional, a length 2 vector detailing the minimum and maximum row to plot}

\item{plot}{whether to plot the graph or return the number of rows to plot}

\item{plotLeftSideOption}{What to display on right axis. Options are: "treatment", "patid" or "both". Default is "treatment"}

\item{xlim}{Range to plot on xaxis. Default is c(-7,60)}

\item{plotCycleLength}{Cycle length is used to add greater highlights to vertical lines. Default is 21}

\item{plotXLegendScale}{What scale to use on xaxis. Options are "days","weeks","months". Default is "days"}
}
\value{
This plot function return the number of row of unique toxicities * patients. This assists in computing optimal size for saved graphs.
}
\description{
This function plots all the toxicities provided over time
}
