% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_eventInfo.R
\name{ToxPlot_eventInfo}
\alias{ToxPlot_eventInfo}
\alias{eventInfo,}
\alias{eventInfo-class}
\title{eventInfo class}
\usage{
ToxPlot_eventInfo(columns, label = columns[1], lwd = 4, col = "grey")
}
\arguments{
\item{columns}{A vector of columns contain the event}

\item{label}{A short name for the event}

\item{lwd}{Line width}

\item{col}{The colour to use for each event}
}
\value{
An object of class eventInfo containing the same slots as paramters taken by ToxPlot_causalityInfo
}
\description{
Stores event data to pass to toxPlot_byToxicity in the causality parameter. The columns containing the data in patientData should hold date of class "Date".
}
