% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxPlot_time.R
\name{toxPlot_time}
\alias{toxPlot_time}
\title{Plot toxicities over time}
\usage{
toxPlot_time(toxDB, patients = character(0), plot = TRUE)
}
\arguments{
\item{toxDB}{Main toxicity database.}

\item{patients}{Build in patient subsetting by a list if not ""}

\item{plot}{TRUE/FALSE plot the graph.}
}
\value{
This plot function return the number of row of unique toxicities * patients. This assists in computing optimal size for saved graphs.
}
\description{
This function plots all the toxicities provided over time
}

