% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_category.R
\name{ToxTable_category}
\alias{ToxTable_category}
\title{Toxicity table of categories by cycle/s}
\usage{
ToxTable_category(rt, cycles = "all")
}
\arguments{
\item{rt}{an object of class robustToxicities}

\item{cycles}{The cycle column names, or index in rt@rt@periodDividerCols of the cycles to tabulate. May also be "all" to use all cycles}
}
\description{
Returns a toxicity table with the requested data according to the ass_TRUE column, for the cycles requested. Note this is a wrapper function which essentially replaces the toxicity names with the categories, updates the toxID's and then calls ToxTable_cycle on the categories. This function could be used as a template for summarising other data.
}
\details{
This function acts as a wrapper for ToxTable_cycle to get category data instead of toxicity level data.
}
