% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_cycle.R
\name{toxTable_cycle}
\alias{toxTable_cycle}
\title{Table: Toxicity worst in cycle}
\usage{
toxTable_cycle(toxDB, cycles)
}
\arguments{
\item{toxDB}{Main toxicity database.}

\item{cycles}{The cycle or cycles to generate the table for}
}
\description{
Generates a table of all toxicities for a certain cycle or cycles. There are options to report all changes in toxicities or just the worst per patient. CTCAE grades and categories can be merged for more sucinct tables.
}

