% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_cycle.R
\name{toxTable_cycle}
\alias{toxTable_cycle}
\title{Table: Toxicity worst in cycle}
\usage{
toxTable_cycle(toxDB, cycles)
}
\arguments{
\item{toxDB}{Main toxicity database.}

\item{cycles}{The cycle or cycles to include toxicities from}
}
\description{
Generates a table of all toxicities for a certain cycle or cycles. There are options to report all changes in toxicities or just the worst per patient. CTCAE grades and categories can be merged for more sucinct tables.
}
\details{
The output relies on the following options:
\itemize{
 \item cycleCycleMerge
 \item cycleColumnMerge
 \item cycleColumnMerge
 \item cycleCategoryMerge
}

 The output also uses the cycleNames in the first column of the table. If cycles are to be merged the name for the first one will be used.
}
\seealso{
\code{\link{print_toxTable_cycle}}
}

