% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxTable_summary.R
\name{ToxTable_summary}
\alias{ToxTable_summary}
\title{Toxicity table summary}
\usage{
ToxTable_summary(rt)
}
\arguments{
\item{rt}{an object of class robustToxicities}
}
\description{
Returns a summary toxicity table with the requested data according to the ass_TRUE column.
}
