% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worstGradeByPatient.R
\name{worstGradeByPatient}
\alias{worstGradeByPatient}
\title{Generate worst grade by patient}
\usage{
worstGradeByPatient(toxDB, colName, df = NULL)
}
\arguments{
\item{toxDB}{an object of class robustToxicities}

\item{colName}{The name of the column created}

\item{df}{(optional) an existing data.frame to add the column to. If noy provided this is created from the data.}
}
\description{
Returns a patient level data.frame containing the worst grade for each patient. The ass_TRUE is used as a filter.
}

